# backend/docs/swagger.yaml
definitions:
...
  api.Response:
    properties:
      data: {}
      error:
        type: string
      language:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
...
  api.SecurityConfig:
    properties:
      max_timestamp_age:
        description: Maximum timestamp age in seconds
        type: integer
      required_headers:
        description: List of required headers
        items:
          type: string
        type: array
      secret_key:
        description: Client-side security secret
        type: string
      security_level:
        description: Required security level
        type: integer
      wasm_enabled:
        description: WASM support enabled
        type: boolean
      web_worker_enabled:
        description: Web worker support enabled
        type: boolean
    type: object
...
  api.UserResponse:
    properties:
      created_at:
        example: "2025-05-01T13:45:00Z"
        format: date-time
        type: string
      email:
        type: string
      id:
        type: integer
      username:
        type: string
    type: object
...
  api.loginUserRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  api.loginUserResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      security_config:
        $ref: '#/definitions/api.SecurityConfig'
      user:
        $ref: '#/definitions/api.UserResponse'
    type: object
  api.logoutRequest:
    properties:
      refresh_token:
        type: string
    type: object
...
  api.refreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  api.refreshTokenResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/api.UserResponse'
    type: object
  api.registerUserRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
  api.registerUserResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      security_config:
        $ref: '#/definitions/api.SecurityConfig'
      user:
        $ref: '#/definitions/api.UserResponse'
    type: object
...
info:
  contact: {}
paths:
...
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return a JWT token
      parameters:
      - description: Login credentials
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/api.loginUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  $ref: '#/definitions/api.loginUserResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: Authentication failed
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/api.Response'
      summary: Login user
      tags:
      - auth
  /api/auth/logout:
    post:
      consumes:
      - application/json
      description: Logout a user by invalidating their access and refresh tokens
      parameters:
      - description: Logout request with optional refresh token
        in: body
        name: logout
        schema:
          $ref: '#/definitions/api.logoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Logout successful
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: Unauthorized if the user is not authenticated
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - ApiKeyAuth: []
      summary: Logout user
      tags:
      - auth
  /api/auth/refresh-token:
    post:
      consumes:
      - application/json
      description: Get a new access token using a refresh token
      parameters:
      - description: Refresh token
        in: body
        name: refresh_token
        required: true
        schema:
          $ref: '#/definitions/api.refreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Token refreshed successfully
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  $ref: '#/definitions/api.refreshTokenResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: Invalid refresh token
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/api.Response'
      summary: Refresh access token
      tags:
      - auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user and return authentication tokens
      parameters:
      - description: Registration information
        in: body
        name: register
        required: true
        schema:
          $ref: '#/definitions/api.registerUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  $ref: '#/definitions/api.registerUserResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/api.Response'
      summary: Register a new user
      tags:
      - auth
...
swagger: "2.0"